load_object: fix buffer overruns.
`real_name` and `obname` are the same size as `name` and `actualname`.
This means that if `lname` is 400 or more characters long, `actualname`
and `obname` will be filled and the `strcat` calls to append `".c"` to
`real_name` and `obname` will result in buffer overruns.

Extend `real_name` and `obname` to make room for `".c"`.

Signed-off-by: Jeremy Sowden <jeremy@azazel.net>