Catch errors thrown when compiling LPC files.
When handling the inclusion of headers, the parser calls the `VALID_READ` apply.
This may fail and throw an exception.  If this happens, the `guard` variable in
`compile_file` is never cleared and the driver will be unable to load any more
objects.

Wrap `yyparse` in a try-catch block in order to clean up properly before
re-throwing.

Signed-off-by: Jeremy Sowden <jeremy@azazel.net>