fix: prevent test_load from corrupting object list
When using `test_load` on a loaded object the object list ends up
corrupted. The details are obscure, but if doing this reproduces the
error:

1. Load any object in memory. Check that both `find_object` and
   `children()` return the single existing object
2. Run `test_load` on the object. Check that `find_object` returns `0`
3. Destroy and create the object. `children()` will now return an array
   with two duplicates of the same object or, sometimes, a `0` and the
   existing copy of the object

In some instances i've seen different objects returned by `children()`
than the one being passed as path (this usually happened with clones).

This PR tries to work around it. I've also reverted #843 because I
believe this PR fixes the underlying cause for the issue, not just the
symptom.

All credit goes to Zoilder (one of the developers on our mud) that
happened to fix this a long time ago. We're just sending these fixes
upstream since we're starting to update our outdated driver.

Please let me know if there's anything I should fix. I've tried to add a
proper test for this, but I don't know if that's the location you're
expecting for it.